+++ src/cc/lib/distributed/graph_engine.cc
@@ -81,9 +81,15 @@ grpc::Status GraphEngineServiceImpl::GetNodeTypes(::grpc::ServerContext *context
             continue;
         }
 
-        const auto index = elem->second;
+        auto index = elem->second;
+        size_t partition_count = m_counts[index];
+        int32_t result = -1;
+        for (size_t partition = 0; partition < partition_count && result == -1; ++partition, ++index)
+        {
+            result = m_partitions[m_partitions_indices[index]].GetNodeType(m_internal_indices[index]);
+        }
         response->add_offsets(curr_offset);
-        response->add_types(m_partitions[m_partitions_indices[index]].GetNodeType(m_internal_indices[index]));
+        response->add_types(result);
     }
 
     return grpc::Status::OK;
@@ -110,13 +116,25 @@ grpc::Status GraphEngineServiceImpl::GetNodeFeatures(::grpc::ServerContext *cont
             continue;
         }
 
-        response->add_offsets(node_offset);
         response->mutable_feature_values()->resize(feature_offset + fv_size);
-        const auto index = internal_id->second;
+        auto index = internal_id->second;
+        size_t partition_count = m_counts[index];
         auto data = reinterpret_cast<uint8_t *>(response->mutable_feature_values()->data());
-        m_partitions[m_partitions_indices[index]].GetNodeFeature(m_internal_indices[index], features,
-                                                                 std::span(data + feature_offset, fv_size));
-        feature_offset += fv_size;
+        bool found = false;
+        for (size_t partition = 0; partition < partition_count && !found; ++partition, ++index)
+        {
+            found = m_partitions[m_partitions_indices[index]].GetNodeFeature(m_internal_indices[index], features,
+                                                                             std::span(data + feature_offset, fv_size));
+        }
+        if (found)
+        {
+            feature_offset += fv_size;
+            response->add_offsets(node_offset);
+        }
+        else
+        {
+            response->mutable_feature_values()->resize(feature_offset);
+        }
     }
 
     return grpc::Status::OK;
@@ -191,9 +209,14 @@ grpc::Status GraphEngineServiceImpl::GetNodeSparseFeatures(::grpc::ServerContext
             continue;
         }
 
-        const auto index = internal_id->second;
-        m_partitions[m_partitions_indices[index]].GetNodeSparseFeature(
-            m_internal_indices[index], features, int64_t(node_offset), dimensions, indices, values);
+        auto index = internal_id->second;
+        size_t partition_count = m_counts[index];
+        bool found = false;
+        for (size_t partition = 0; partition < partition_count && !found; ++partition, ++index)
+        {
+            found = m_partitions[m_partitions_indices[index]].GetNodeSparseFeature(
+                m_internal_indices[index], features, int64_t(node_offset), dimensions, indices, values);
+        }
     }
 
     for (size_t i = 0; i < features.size(); ++i)
@@ -275,10 +298,15 @@ grpc::Status GraphEngineServiceImpl::GetNodeStringFeatures(::grpc::ServerContext
             continue;
         }
 
-        const auto index = internal_id->second;
-        m_partitions[m_partitions_indices[index]].GetNodeStringFeature(
-            m_internal_indices[index], features, dimensions.subspan(features_size * node_offset, features_size),
-            values);
+        auto index = internal_id->second;
+        size_t partition_count = m_counts[index];
+        bool found = false;
+        for (size_t partition = 0; partition < partition_count && !found; ++partition, ++index)
+        {
+            found = m_partitions[m_partitions_indices[index]].GetNodeStringFeature(
+                m_internal_indices[index], features, dimensions.subspan(features_size * node_offset, features_size),
+                values);
+        }
     }
 
     response->mutable_values()->append(std::begin(values), std::end(values));